requirements.txt
feedparser==6.0.10
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
openai==1.3.5
sqlalchemy==2.0.23
beautifulsoup4==4.12.2